auto generate discord invite link/bot refactor
- Created new invite_link attribute for DiscordGuild models
- Within on_presence_update, the bot will now create a new invite link and save it to the DiscordGuild's invite_link attribute, IF the attribute is empty. It will not create a new link fi the field is populated
- Refactored raptorbot_util into separate files based on their function
- Discord Button on website will not appear if there is no DiscordGuild
