+++ raptorWeb/raptorbot/models.py
@@ -77,6 +77,12 @@ class DiscordGuild(models.Model):
         guild_id = models.BigIntegerField(
             verbose_name="Guild ID",
             default="None")
+
+        invite_link = models.URLField(
+            blank=True,
+            verbose_name="Discord Invite Link",
+            help_text="The invite link to your Discord Server. This is automatically generated by the Discord Bot"
+        )
         
         total_members = models.IntegerField(
             verbose_name="Total Members",
@@ -87,9 +93,6 @@ class DiscordGuild(models.Model):
             default="None")
 
         def __str__(self) -> str:
-            return f'{self.online_members}/{self.total_members}'
-
-        def get_guild_name(self):
             return self.guild_name
 
         def get_guild_id(self):
@@ -103,4 +106,4 @@ class DiscordGuild(models.Model):
 
         class Meta:
             verbose_name = "Discord Guild"
-            verbose_name_plural = "Discord Guilds"
+            verbose_name_plural = "Discord Guild"
