+++ raptorWeb/templates/raptormc/base.html
@@ -101,14 +101,12 @@
           <div class="container">
             <div class="navbar-header d-flex">
               <div class="row p-2 pt-1 pb-1">
-                <a class="col-3 col-sm-2 col-lg-1 btn btn-outline-secondary" role="button" data-bs-toggle="tooltip" 
                 {% if discord_guild %}
-                data-bs-title="{{ discord_guild.online_members }}/{{ discord_guild.total_members }} Members Online" href="https://discord.gg/GEhyhku" target="_blank">
-                {% else %}
-                data-bs-title="No Guld" href="https://discord.com" target="_blank">
-                {% endif %}
+                <a class="col-3 col-sm-2 col-lg-1 btn btn-outline-secondary" role="button" data-bs-toggle="tooltip" 
+                data-bs-title="{{ discord_guild.online_members }}/{{ discord_guild.total_members }} Members Online" href="{{ discord_guild.invite_link }}" target="_blank">
                   <img id="discordBanner" src="{% static 'image/discord_banner.webp' %}" alt="Discord">
                 </a>
+                {% endif %}
               </div>
             </div>
           </div>
